Fixed wrong migration for resources without volumes
This prevents NPE when the controller starts up, but unfortunately
a previous migration already deleted the DRBD node ids of the resources.
Those ids are unrecoverably gone.

This commit has to assign new node ids to the resources. If those
ids do not match the possibly already (and still) running DRBD,
DRBD will be very unhappy about it and the user will have to manually
disconnect and re-connect the DRBD resources.
Linstor's "drbd adjust ..." is not enough!