Set access context for REST API calls to the public context
A system context was arbitrarily injected for API calls.
However, the system context cannot normally sign in, and
objects created in the system context are not normally
modifyable by unprivileged sessions.

Any API that is called from a session that did not perform a
sign in before (e.g., by calling the SignIn API) must run in
the anonymous public context.
Therefore, changed the injected context to @PublicContext.