Fixed volume-definitions not being deleted on satellite
Until now, only the AbsStorageProvider called the notifyVolumeDeleted
method of the deviceManager. This triggered sending the notification event
to the controller and also actually deleting the satellite-local
volume-definition.

However, if a volume-definition was deleted of a diskless resource
the AbsStorageProvider was not involved and thus was not able to call the
notifyVolumeDeleted method.

This fix simply moves the call of notifyVolumeDelete method into the
DeviceHandlerImpl, which is executed in every case regardless if diskful
or diskless.