Fixed race-condition when creating vlmDfn too fast
Having a resource deployed, creating a new volume-definition triggers the
update-satellite cycle as expected, which will finish with stlt->ctrl
notifyResourceApplied containing all layer-specific data for every
volumes within the resource.

If in the meantime an additional volume-definition was created, the
controller might have created the (controller-local) second volume of
that resource. Later, when processing the notifyResourceApplied payload,
the controller iterates over its own volumes and looks for matching
volume-data within the payload. As the notifyResourceApplied was started
without the second volume(-definition), it will be missing the data for
the controller-local second volume, resulting in a NullPointerException.