Added all storPools to prioProps for NVMe's PrefNic
NVMe needs an IP address on resource level. As the PrefNic property can
conviniently be stored on storage pool level, the NVMe layer has to
decide which PrefNic to take (in case there are multiple defined).

This patch simply asks all volumes storage pools (in order of volume
numbers) for the possibly set PrefNic. If no storage pool or volume
has set the property, resource, node and satellite are asked