Remove DRBD dependency from DISKLESS provider
This dependency can be savely removed as the satellite might be able to
create a diskless storage-pool, but the actual DRBD-resource will be
rejected by the controller as the DRBD-layer is not supported.

Removing this dependency is necessary to be able to set properties on
diskless storage pool where DRBD is not installed (i.e. NVME, STORAGE-only
etc.).