Fixed resize and detailed size-reporting issue
When the ZfsProvider updated the internal state based on the output of
zfs list, the previously set usable size (which was the expected size in
case of a resize-event) was overridden.

Additionally, when the satellite reported the usable size back of LVM
devices to the controller, the data was always outdated.