Fixed Exception when reconnecting to peer
A satellite was added twice into our internal ping-list and therefor later
also to the reconnect-list. When reconnecting (both) the first succeeds,
everything is fine, but when the second also succeeds, that second Auth-
API call answers the on the first connection "OTHER_CONTROLLER" and closes
the connection. That first connection still had the full-sync in progress
which caused a PeerNotConnectedException due to the closed connection.