DevMgr: Call forget-peer when remote-rsc is deleted
If a remote resource is deleted, linstor now resets
the bitmap to day0 using drbdsetup del-peer and forget-peer
commands.

This also means that the StltAPI no longer removes the
remote resources, but allow them to go through the
DeviceManager. However, the DeviceManager still iterates
only over local resources, but via the resource-definition
the remote-resources which have the DELETE flag set are
still reachable.

This is part of solution for a reproducible splitbrain bug

* Create 2 linstor resources
* Use the drbd (primary, secondary is enough)
* Delete one of the linstor resources
* Recreate the just deleted resource

 -> splitbrain