Partially reverted "LvsUtils now parses json output for 'lvs' and 'vgs'"
Reverted parsing json output introduced in
4c640684696719b3d9c6eb45fe373e0332563ea9. That commit also introduced
a potential bugfix, that LVs with the same name but in different VGs
cannot be differentiated.

This was not that big of an issue, as linstor maps the resource-name
in the LV name, which means that linstor cannot create same named
LVs in different VGs (storage-pools).

Even if there exists non-linstor-managed LVs with linstor-managed LVs
sharing the same name, linstor usually only queries the LVs of linstor-
managed VGs.

That means, this bug could only occur, when a user gives linstor at least
two VG (A and B), manually creates an A/res_00000 and afterwards tries to
create a linstor-managed resource B/res_00000. Due the now fixed bug,
linstor would have taken the A/res_00000 instead of actually creating the
LV in B.