LayerData: VlmDfn now supports mulitple instances of same kind
(copy pasted from previous commit)

Volume definitions stored the layer-data in a map with DeviceLayerKind
as key. This led to the problem that only one layer data of the same
type could be stored. Up until now, this was also sufficient, although
not intended.

The current version has basically a map where the key is a pair of
DeviceLayerKind and a String representing the resource-name-suffix.