DrbdLayer: added support for external metadata
This patch also allows layer specific volumeData to be null.

For example, if a drbd resource has a volume with internal
metadata an a second volume with external metadata, two rscData
objects will be generated (with the suffixes "" and ".meta")
"" will be the usual data-resource (here we simply assume that is
directly the storage resource). this data-storage-resouce is filled
for both, the volume with the internal and with the external metadata.

The ".meta" storage rscData will only have one vlmData, namely for the
second volume, which has external metadata. The other volume-number of
that storage rscData will return a null object.

This was needed to reduce the necessary communication between layers
as much as possible.