DB: Fixed rscDfn and vlmDfn layer data loading
Layer data for resource-definitions and volume-definitions were only
loaded when needed by a resource or volume.

Resource-definitions which already had allocated layer-specific data
(like TCP-port or minor number for DRBD), but had no existing resources
were simply not loaded in the startup of the controller.
That had the result that the allocation was not tracked by linstor when
the controller was restarted and therefore already allocated numbers
were reused by new resource-definitions and volume-definitions, resulting
in unique constraint violation exceptions.