LvmThin: Fixed incorrectly displayed allocated size
There is still another issue, namely that although "r lv" fetches the
most upTotDate thin sizes from the volumes, these updates are only based
on volumes, not the actual volume-layer-data. That means that we do have
the allocated size for the volume, but we dont know which layer consumes
how much of that size.

To fix this, we need to make the fetching based on volume-layer-data and
call LayerRscDataMerger.restoreRscData (which performs a merge of a given
rsc-layer-api-data to the actual rsc-layer-data)