Eclipse: Workaround
There seems to be a bug in the eclipse compiler. If this patch does not
explicitly cast the lambda to ExceptionThrowingSupplier, the compiler has
to decice if this is a *Supplier or a *Runnable (by ignoring the return
type).
The compiler "could" figure out that both are possible and one if them
is required from the caller. This was working in previous versions of
eclipse, in 4.12.0.20190614-1200 this seems to be broken.