Drbd: ExtMetaStorPool is now persisted to DB
This patch fixes the bug where chaning the controller property for
external-metadata-storage-pool after a resource was created actually
changed the target storage pool on the satellite for already existing
drbd resources.

Now that property is copied and stored in the database. Once initialized
the external storage pool cannot be changed.

Current limitation: "cannot be changed" also applies to toggling to
diskless and back. However, toggle disk is currently not supported
for volumes with multiple storage pools