GenericDbDriver: Added check for missing ';'
If an SQL statement was passed to GenericDbDriver.runSql, but was missing
the final ';', the statement was not executed and silently ignored.
This patch will try to execute that final statement. If the statement was
not ment to be executed (some strange comment or other instruction), the
jdbc driver will respond with an SQLException anyways.