AbsStorageProvider: Added check for size of created volumes
fallocate apparently does not fail if it could not fully allocate the
wanted size. Instead it returns 0 and continues without message while
the resulting file shows 0 allocated bytes.

Therefore we simply perform a check if the size of the just created
volume is >= as the expected size in case of thick provisioned
volumes, and at least > 0 in case of thin provisioned