Nvme: VlmData are now set to exists
When an initiator connected successfully or a target was created
successfully, the nvme-resource-data was set to exists=true
but not the nvme-volme-data. This caused for example the SysFsHandler
to not set the throttling settings on that device.
(This effect was only visible for initiator volumes as the target volumes
had their backing device which also got the throttling set)