Moved core objects to dedicated package
The package com.linbit.linstor got quite bloated. This patch moves the
core objects (node, resource, resource-definition, ...) into the package
com.linbit.linstor.core.objects.

In the server project this effects the corresponding interfaces,
the *Data classes and some generic (controller + satallite) factories

In the controller project this effects the database drivers as well
as the controller specific factories

In the satellite project this only effects the satellite specific factories