DbDriver: Added Database constants generator
The generator needs to be run after flyway migration is finished.
The generator only uses JDBC functionality to gather tables, columns,
types and primary constraints (addtional constraints would be foreign
and unique constraints)

The generator returns the rendered java code as String, which has to be
written by the caller to the desired path. (rendered package and classname
are configurable)