SysFsHandler: cleanup caches on delete
When a volume is deleted, the caches of the previously set values
was not cleaned, which prevents (re-)setting the same
"maj:min value" property even for different (deleted and re-created)
devices.

In other words, setting such SysFs properties on i.e. storage pool,
the first volume will get that property set as expected. When deleting
and re-creating the same resource (thus a volume with the same name),
might get the same maj:min as before. If the caches were not cleared
during the re-creation, the second volume will not get the sysfs
property set.