NvmeUtils: Fixed potential bug with exists flag
This is a partial revert of e3882830bfc521c9d5136aa03960e05577c18993

If linstor for some reasons have a different state than the system (i.e.
linstor thinks an nvme-volume exists but it does not, or the other way
around), the exists flag was not updated properly.

I.e. in the first case when creating a namespace. If linstor thinks
that the namespace should be created, but the namespace already exists
(which can easily happen by i.e. restarting the satellite), prior this
patch linstor would not have set the nvmeVlm to exists=true.