LvmProvider: changed "wiping_in_progress" lv-suffix
When choosing long resource names (max 48), which gets extended by
UUID and "_linstor_wiping_in_progress" when the LV is wiped before
deleted, LVM internally also extends the name of the volume-group
into its internal LV-name. LVM has an internal limit of the length
of LV names (currently 128 including trailing '\0'), which can be
exceeded with such a '$uuid_linstor_wiping_in_progress' extension.

This patch removes the additional uuid, and uses a '-' as connecting
character, which cannot conflict with snapshot names, as they connect
with '_' to the leading "$rscName_$vlmNr" naming schema.