ETCD: fixed duplicated key in txn bug
ETCD does not allow multiple requests for the same key in the same
transaction.

Linstor transaction-manager already filters the requests so that only
the last request for a given key (exact match) will be sent to the
etcd server.

However, two operations, node lost and toggle-disk, caused an update to
key A/B and a deletion of A/* (or first the delete and afterwards the put).
This also resulted in the same issue.

In order to fix this bug, the ETCD drivers for node and storage-volumes
were changed so that they do not issue a delete of A/* but a list of
deletions of A/B, A/C, ... so that the already existing cleanup-mechanism
of linstor's transaction-manager gets rid of the unnecessary requests.