RetryTask: Cleanup successfully deleted resource
If a resource-operation returns an error from the satellite, the Flux
will be canceled. The problem is, that only that Flux knew the next steps.

That means: when the user issues "linstor r d .." and the satellite
fails to delete that resource, the RetryTask will retry the deletion as
expected. However, if retrying the deletion become successful, the
RetryTask has no information about what to do next (i.e. delete the
resource from the database), as this information was only stored in the
original Flux.

Registering the "next steps" of the original Flux along with the just
failed resource might cause duplicate deletions / modifications of i.e.
ResourceDefinition objects (i.e. "rd d" would register all resources
"in case you fail, retry and then try to delete the resource-definition".
If more than one fail, and more then one succeeds during a later retry-
attempt, all of the success full "resource deletes" will try to
delete the same resource-definition.)
This would require large rewriting of internal API handling.

The current solution is not perfect, as the RetryTask still has no idea
about the "next steps", but only cleans up the database if the resource
had its DELETE flag set.
That means, if we have 2 resources, one diskfull and one diskless and
we try to delete the whole resource-definition and the deletion of the
diskless fails - the Flux gets lost. The RetryTask will retry, and if it
succeeds it will also delete the resource. However, it will NOT trigger
the (originally expected) deletion of the diskfull resource and also
skip the deletion of the whole resource-definition.