RetryTask: continues previous flux
This patch modifies the RetryTask in a way that it can update
the satellite, and afterwards continue the flux stream as if
the operation never failed.

This operation includes a "rd d .." where i.e. the first of
multiple resources failed. This patch not only retries the deletion
of the resource, but also continues the deletion of the other
resources of the originally specified resource-definition as well
as deletes the resouce-definition itself.

1b37baac mentions the issue of a large rewrite of API handlers.
Turns out that the mentioned resource definition delete
flux already has such a check. Other flux steps should only
consider their own resource.