FileThinProvider: fixed loop_device_mapping output
The DeviceProviderMapper did not include the FileThinProvider which is used
for iterating when cleaning up. The cleanup method renders the content for
/var/lib/linstor/loop_device_mapping.

Another bug was that the two FileProvider and FileThinProvider somehow
raced each other, always overriding the same loop_device_mapping.
Actually, they still do, but now the source of the content is a shared
(static) map, instead of two different provider-local maps which only
"synchronize" themself through calling `losetup` in each prepare step.