DrbdLayer: Fixing del-peer race condition
Copied from java-comments:

Race condition:
If two linstor-resources are deleted concurrently, and one is much
faster than the other, the slower will get an "unknown connection"
from the drbd-utils when executing the del-peer command.
In that case, we will still try the forget-peer.
If the forget-peer command succeeds, ignore the exception of the failed
del-peer command.
If the forget-peer command also failed we ignore that exception and
re-throw the del-peer's exception as there could be a different reason
for the del-peer to have failed than this race-condition