Database: Fixed incorrect linking
It was possible that an object (node, resource, ...) was created with one
database driver, but after a controller-restart loaded with a different
one.

This does not mean that we had a mixup between ETCD and SQL (otherwise
the prevsiously stored data would not have been found). If some
special ETCD drivers (like NodeEtcdDriver) stored an object,
the DatabaseLoader had requested "any" driver implementing the correct
interface, which might have been a more general driver (i.e. not the
special NodeEtcdDriver). This caused some issues as there are good
reasons for the specialized drivers to exist.

This patch introduces additionally a bunch of new interfaces, basically
only extending the general *DatabaseDriver (which include persist, update
and delete methods) with a *CtrlDatabaseDriver whereas only the latter
interface includes load methods.