AutoTiebreaker: Allow takeover
This patch allows the execute a "resource create .." even if the target
resource exists AND the target resource is a tiebreaker.

If the new resource should be diskless, the API call does nothing but
remove the TIE_BREAKER flag from the resource and updates all
related satellites.
If the new resource should eb diskful, beside removing the TIE_BREAKER
flag a toggle-disk is executed.