NumberAlloc: Removed LoopGuard
The loop guard could only be triggered by implementation errors
in the code, but not by invalid data supplied to the methods.
Even with an unsorted input array, the various search loops
are guaranteed to end after a limited number of iterations,
because the search range is shrinked by each iteration.