DRBD: additionally copy .res files to backup dir
At the end of the DrbdLayer.process method, the .res file is copied into
/var/lib/linstor.d/.backup/

There are two reasons for this patch:
1) if the user somehow starts the satellite without a proper LS_KEEP_RES
configured and all .res files are accidentially gone (that could lead to
a HA controller not being able to start any more)

2) the backup of the .res file is done as the final step in the DRBD
layer. That means, if a drbd command fails, we now have the .res file
which caused the failure, but also the version of the .res file which
worked before the failure. This might come handy when looking for the
root cause.

Obviously this is not a "restore from backup and everything works again"
type of solution, as for example when toggling diskful to diskless the
backuped .res file still contains the path to the storage device, although
that device was deleted in the meantime.

The .res file is eventually deleted from the .backup directory when
linstor successfully deleted the actual drbd-resource. Otherwise the
.backup directory never gets cleaned.