LvmProvider: disabled unimplemented async wiping
Renaming the LV and leaving behind if the actual deletion fails
caused some issues. As the async wiping is not even implemented
we simply disable this feature for now.

Now the LvmProvider directly tries to remove the LV. If that fails
a StorageException aborts the whole resource-delete process, which
means the resource will be added to the controller's retry task.