Replaced some external ls/rm commands with Java code
Listing, sorting and unlinking files or directories using Java
code is supposed to be portable and also more robust than
using system utilities, due to a lack of standardization
of the utilities' parameters and output.

There is still a lot of potential for improvement here, many
of the file operations could be more robust (e.g., cases where
it attempts to delete something that doesn't exist in the first
place could be detected and handled gracefully).