Drbd,Nvme: Added support for snapshot
Such a setup is quite special. Node A as the diskful DRBD is based on an
NVMe initiator (that means there is no local disk), whereas the other node
B does not have a running DRBD, but provides the actual storage as an
NVMe target.
That means, node A (where DRBD is actually running) has to execute a
drbdadm suspend-io $rscName
and once that finished, node B has to take the snapshot.
After the snapshot was taken node A must afterwards resume-io.