VlmDfn: Added GROSS_SIZE flag
If this flag is set, the storage layer will first allocate the
specified block with exactly the size of the volume-definition.
Every layer will then reserve a bit of the available space reducing
the resulting usable size.

Examples of a 100G vlmDfn with internal DRDB metadata:

Without --gross:
results in 100.02G LV with exactly 100G usable space.

With --gross:
results in an exactly 100G LV with a usable size of 99.98G