Diskless: Fixing unnecessarily needed storagepools
If a DRBD or a Writecache layer is configure above an nvme,
the target resource will not create the LV for external metadata or
writecache. However, until now, the target node also had to have
the storagepool definied in the corresponding property
("StorPoolNameDrbdMeta" / "Writecache/PoolName"). Otherwise
the controller reported an error which would have led to a
nullpointer.