Nvme: Target shares drbd-node-id with its initiator
Following scenario:
create nvme-target
create nvme-initiator
create drbd on top of the initiator
  this also creates meta-data basically on the target (through nvme)
use the drbd-device

destroy drbd-device (only on initiator, data will stay on target)
disconnect nvme-initiator from target

re-create nvme-initiator and drbd
  (possibly on a different node then before)

The new nvme-initiator has to know the node-id which is already
contained in the meta-data on the nvme-target.

This is basically done by allowing 2 linstor-drbd-resource-data to share
the same node-id. When the linstor-drbd-resource is created above the
linstor-nvme-initiator, linstor simply looks up the target-resource and
copies the stored node-id.