ETCD: added deadline (timeout) for etcd requests
backoffRetry will retry the request for a long time.
deadline(...) limits the maximum waiting time.

DO NOT USE sync() here, as there appears to be a bug in the grpc
implementation (consuming a runnable and retrying the now null
object instead of the runnable which causes the sync() to wait
without even trying)