Revert "Partial revert of ADD COLUMN AFTER clause"
This reverts commit cd3a892c45eb65f4bbef63489271f56b3ea1150b.

We do need the correct order of primary keys in the
GeneratedDatabaseTables.java
That class is created while reading the columns in order from
the database. If the snapshot_name is added as last column,
the ETCD engine will try to parse the
"<resourceName>:<snapshotName>:<volumeNumber>" for a volume
as
"<resourceName>:<volumeNumber>:<snapshotName>" which will  obviously
fail.