Drbd: fixed resize while resync uncleard flag
If DRBD is currently resyncing a resource and the user issues a resize
to the volume, that action will fail.
Failed resource-actions will be added to the retry-task, which should
continue possible next flux-steps (or cleanup-fluxes).
However, due to NotifyResourceFailed the "next step" was lost (not set
in the flux step setting DRBD_RESIZE flag, which means that failed
action got registered without the next step, and nobody was triggered
afterwards to properly set the next step).