VlmDfn: Changed method of toggling GROSS_SIZE
If a volume-definition was created with enabled GROSS_SIZE, but then
was resized (without re-setting GROSS_SIZE) the GROSS_SIZE flag was
disabled (because it was not set during the resize command).

Now the GROSS_SIZE will stay the same if no flags are set.
If you want to remove the existing GROSS_SIZE, the client has to sent a
-GROSS_SIZE (i.e. a "negative" GROSS_SIZE flag).