docker,controller: add client for debugging
Adding the linstor-client to the image was requested as it makes
debugging easier.

This is for debugging only. As this pulls the client from the repo, this
usually will fetch an older client (assuming a release cycle where all
of LINSTOR is released). So if there is a breaking change between
controller and client, one would need to sync the client to the repos
first, and then build the controller image. But again, even if it
breaks, it is fine, this is not the official client image. And as the
repo is still configured in the controller images, eventually the one
debugging can updated to the newest client within the container via
'yum'.