Db,Flyway: do not ignore future migrations
If a migration failed and the user only reverts the linstor-server
version but forgets about rolling back the database, linstor looks
like it works well, until it starts to run into weird errors.

This patch causes Flyway to throw an exception when it detects
a migration applied by a higher version.