containers,satellite: allow to use host LVM
While I think this should not be used in general, it gives the satellite
container the possibility to use the host's lvm. Again, I would not use
it in general, but we have seen host/container combinations where LVs
don't show up on the host (while perfectly fine in the container).
Usually this can be fixed by running the usual scan/change commands on
the host.

The whole thing is triggered by bind-mounting the host's /proc to
/host/proc. This then makes the container a "super privileged
container"[1] for all LVM commands.

It also adds util-linux (nsenter) as an explicit dependency.

[1] https://developers.redhat.com/blog/2014/11/06/introducing-a-super-privileged-container-concept/