Lvm: added --config option to filter required pvs
Every lvm-related command now specifies a
--config="devices { filter = ['a|/dev/...|'|, ..., 'r|.*|'] }"
where the filter is accepted list is build using an unfiltered
pvdisplay. The filter is cached for every collection of volume-groups.

If any lvm command fails, first the specific cache entry is rebuild
(and recached) and the command is retried once.
If the second attempt also fails, the exception is thrown as usual.