Autoplacer: Introducing new Autoplacer
NOTE: TESTS ARE FAILING

This new approach of the autoplacer works in 4 steps:

1) Filter all usable storage pools
A storage pool is considered "usable" if
* the user has enough rights
* the storage pool is NOT diskless
* the storage pool has enough free space
* the storage pool matches the filtered name (if given)
* the storage pool matches any of the filtered provider kind (if given)
* the node supports all of the required layers
* the node of the storage pool matches the filtered name (if given)
* the node does not have filtered resources deployed (if given)
* the node's properties matches to same/different filters (if given)
  (in this step only fixed value property-filters like "key=value"
   are considered)

2) The resulting list from step 1 will be rated by different strategies
Currently only a free-space stragety is implemented but we can
add more if needed.

A strategy can return an arbitrary double value. Those values get
normalized (per strategy) giving the "best" storage pool a rating of
1.0

Afterwards, although not fully implemented yet, every strategy can
be weighted.

The normalized and weighted ratings of each strategy will be summed
up as a resulting score of every given storage pool.

3) An optional user-defined JavaScript can manipulate those scores
This optional step can override the scores from the previously
calculated strategies, basically allowing the user to implement
their own strategy. It is their decision whether they simply
add their score to the existing or completely override the other
scores.

4) Selecting the actual set of storage pools
Starting with the highest rated storage pool, all not yet considered
filters (like key-only property filter) are evaluated.
The first selected storage pools might exclude other storage pools
which means that the highest rated storage pool might have to be
matched with the worst storage pools to meet the place-count criteria.
As the resulting set of storage pools might not be the best possible,
the selection continues searching for other candidates as long
as the highest possible sum of score of the remaining storage pool
is able to exceed the currently best selection.