Autoplacer,API: Added stor_pool_name_list
In addition to the existing stor_pool_name we now also have a list
of the same.

If the list attribute is set and is not empty, the list will be taken
and the content of the now deprecated stor_pool_name is ignored.
Otherwise the stor_pool_name will be added as only element to the
list for further processing.