RscCrt: Reworked ResourceCreateCheck
The previous version kept the state of the last checked resource-
definition, meaning that if the previous resource-definition
had already an nvme-target for example, the now checked resource-
definition was not allowed to have a "second" nvme target, although
the new resource-definition is just getting its first resource.

The current implementation has only the apiCtx as a class-variable
meaning no state is stored which could influence the next check.