Autoplacer: Redirected qmvs to new autoplacer
In order to keep compatibility with the current qmvs API
the qmvs now iterates through all related storage pool names
and runs a filtered autoplace-run for each storage pool name.

This was necessary as the qmvs (including its API) is strongly
dependend on this behaviour of grouped storage pool names.