Fix: thrown exc in finally block may suppress original exception
When an exception is thrown in the try or the catch block, an additionally
thrown exception from the finally block will override the origianl exception.
The original exception is NOT automatically added as suppressed exception.

This patch modifies all finally blocks that may suppress an exception in a way
that the old exception is added as suppressed exception into the new exception
from the finally block.