ParseThinSpace: Fixed NPE due to a race condition
This race condition is quite hard to trigger as it requires
something like a "storage-pool list" that triggers the
fetching of thin spaces, but during the satellite already
authenticated but not yet processed its full-sync. That means
the satellite does not know about any of its storage pools
and therefore returns an empty list.

The ParseThinSpace method simply does not expect the list
to be empty, as at least the DfltDisklessStorPool should be included
(that storage pool cannot be deleted)