Autoplacer: Fixed "autoplaced on 0 nodes" bug
Copied from comments in new test case:

Scenario: The Autoplacer.Selector happily finds the first candidate but is
unsure if that is the best candidate or not, so it continues searching for
a while.
When the selector is confident enough that there cannot be any better
candidates it simply returns the stored candidate.

The bug was that the returned "stored" set of candidates was actually the same
reference as the Selector's internal "currentCandidateSet", which gets cleared
if a new search starts (i.e. "the search continues" after a candidate was found)