TcpConnectorService: Attempt resolving deadlock
(Mostly copied from java-comment)

the selectNow has to be inside the synchronized block as otherwise
it would be possible that the TcpConnector thread is already past
the previous .select(), THEN another thread (i.e. reconnector)
calls .wakeup() which is immediately consumed by the .selectNow()
BEFORE the thread calling .wakeup() could register some new
listeners. We suspect that we ended up in a deadlock occasionally
because of this.