Fixed some tests
If the tests were executed in the wrong order, the new H2ErrorReporter
finds the same database as used by other tests. Although the
DBLoggingTest starts with deleting the H2ErrorReporters database-
file, H2 still finds the entries and is very confused that
there are 9 or 43 error reports instead of the expected 2 which
are generated by the DBLoggingTest.

This patch simply calls a "dataSource.close()" at the end of
the necessary tests, Controller and Satellite.

Only 2 TestExtCmd tests remain where such a .close() call
is not streight forward to implement. However, the ExtCmd
only uses the errorReporter for ".logTrace" and alike, no actual
ErrorReports are created.
Until that stays this way, these "unfixed" 2 TestExtCmd tests
should not be a problem.