SnapshotShipping: Generated name can now skip ids
By default snapshot-shipping generates a "ship$id" snapshot name.
If the user for example already created a "ship0" snapshot,
the controller will now find that snapshot, and will continue with
"ship1" instead of complaining about naming-conflicts"

The snapshotname of the previous shipping is stored on the
resource-connection, so skipping a snapshotname for shipping
should not cause a problem.