SnapshotShipping: Added inactive resource-flag
A resource with this flag will not return any device-path for
its volumes (similar as an nvme-target). Additionally the
satellite also will skip (or even destruct) all layers above the
storage layer, so the device will not be held open by any other
layer.

When the flag is removed the satellite will also re-create
the configured layers.