SnapShip: Prevent multiple merging of snapshots
(mostly copied from java-comment)

Skip merging the snapshot when the snapshot does not exist.
This might happen when we already merged the previous shipping, but
before the SHIPPING_TARGET flag is removed from the snapshot, this
satellite receives a new update.
The snapshot is gone, the merge was successful (hopefully :) ) but
the SHIPPING_TARGET is still present.

In this case, we should skip the finishShipReceive step, as that
might delete the original volume (which would be very bad in this
case as we also no longer have the snapshot which should usually
replace the original volume, so we would end up with no data at
all for this linstor-volume)