DevMgr: Improved send ready-state handler
Before this patch, only the topmost layer got called to send its
"resource ready" event. However, at least in case of DRBD, if there
was no DRBD-device (i.e. when DRBD was over an nvme-target), then
the DRBD layer cannot send its ready-state.

With this patch the called layer has to return a boolean whether
or not it sent a ready-event to the controller. If the current
layer did not send anything, the next layer will be called.