AbsStorProvider: Fix snapshot deletion bug
An old idea of snapshot-shipping was to merge the snapshot into
the volume when the snapshot-shipping is finished and the controller
marks the snapshot for deletion. That idea required the origin to
still exist.

However, we do have usual cases where the user already deleted the
resource and wants to cleanup existing snapshots - this scenario
has nothing to do with snapshot-shipping.

The merge was also implemented differently in the meantime
which makes this check completely unecessary.