Stlt,DevHandler: Fixed bug when deleting an inactive resource
Although the satellite correctly deletes the layers of the resource,
the resource is never marked to notify the controller about deletion.
This had also the effect that the satellite did not properly call
the .delete() method of the result after deleting it.

Recreating the resource later will cause a DivergentUuidsException.