AutoDiskfulAllowCleanup: Remove excessive resource
DrbdOptions/auto-diskful:
(recap)

If a diskless resource is primary for X minutes (whereas X is the
vlaue of the property), Linstor will automatically perform a
toggle-disk. The used storage-pool is found using the auto-placer
with the resource-group's auto-place config, but limited to the
speicifc node obviously.

New behavior:

After the toggle-disk Linstor waits for DRBD to sync up.
Afterwards Linstor tries to remove an unnecessary resource, as the
count of diskful replicas might have exceeded the configured
--place-count of the resource-group's auto-config.

The resource to be deleted is chosen by a part of the autoplacer.
The second step of the autoplacer rates all storage pool, giving
them specifc scores (based on free space, reserved thin-size and
deployed resource count), and tries to choose the storage pools
with the highest score for autoplacement.

When deleting the excess resource after auto-diskful, the score
of the already used storage pools is re-evaluated, and the resource
using the storage pool with the lowest score will be removed.
The idea is to increase that storage pool's score for the next
autoplacement.

DrbdOptions/auto-diskful-allow-cleanup:

If this property is set to false, Linstor will not consider this
resource for deletion during such a auto-diskful-cleanup