containers: add pkgcache
This is useful on very slow architectures (e.g., s390x in qemu). There
we can use a pkgcache with noarch pkgs built on a fast builder (i.e.,
amd64).

We could even use this cache on amd64 on the first build, if we schedule
the docker build after the other regular pkg builds in lbbuild.
Obviously we would have to wait till we can fill the cache. All over I
think not having the cache, but building docker in parallel is faster.

Building in parallel (i.e., docker and rhel7.0) might introduce a
concurrency race on amd64. Imagine rhel7.0 already moved the
linstor-common pkg to the cache, but not the satellite/controller ones
when the docker build checks the cache. We can avoid this by not copying
"*" in the Dockerfiles, but using a sub-shell and copying all the
specific pkgs we need. Then this essentially becomes an atomic
operation.

So from now on one wants to build releases as:
$ lbbuildctl build linstor-server --arch=amd64 ... # which populates the
cache, and then
$ lbbuildctl build linstor-server --arch=s390x -d docker ... # to build
the container image on the slow arch with the cache.

In "lbbuild mode" the build system makes sure the pkgcache/ exists and
has at least 1 file, but not necessarily pre-built rpms.