Stlt: Call systemd-notify READY=1 after first fullsync
The call "systemd-notify READY=1" is only made if the environment-
variable NOTIFY_SOCKET is set and not empty

For now we are not changing the default service-type from
"simple" to "notify". When changing to "notify" be aware that
systemd will shutdown the satellite if the fullsync does not happen
within a timeout. You might want to disable that timeout by adding
  TimeoutStartSec="infinity"
to the service file.
However, if you manually want to start the service with i.e.
  systemctl start linstor-satellite
that call will block until the satellite receives a fullsync.